test split for iv
