+++ pyfixest/estimation/feiv_.py
@@ -143,6 +143,8 @@ class Feiv(Feols):
         store_data: bool = True,
         copy_data: bool = True,
         lean: bool = False,
+        sample_split_var: Optional[str] = None,
+        sample_split_value: Optional[Union[str, int]] = None,
     ) -> None:
         super().__init__(
             FixestFormula,
@@ -159,6 +161,8 @@ class Feiv(Feols):
             store_data,
             copy_data,
             lean,
+            sample_split_var,
+            sample_split_value,
         )
 
         self._is_iv = True
