+++ tests/test_errors.py
@@ -9,7 +9,6 @@ from pyfixest.errors import (
     EndogVarsAsCovarsError,
     FeatureDeprecationError,
     InstrumentsAsCovarsError,
-    MultiEstNotSupportedError,
     NanInClusterVarError,
     UnderDeterminedIVError,
     VcovTypeNotSupportedError,
@@ -119,11 +118,11 @@ def test_iv_errors():
     with pytest.raises(NotImplementedError):
         feols(fml="Y ~ 1 | Z1 ~ X1 ", data=data).wildboottest(param="Z1", reps=999)
     # multi estimation error
-    with pytest.raises(MultiEstNotSupportedError):
+    with pytest.raises(NotImplementedError):
         feols(fml="Y + Y2 ~ 1 | Z1 ~ X1 ", data=data)
-    with pytest.raises(MultiEstNotSupportedError):
+    with pytest.raises(NotImplementedError):
         feols(fml="Y  ~ 1 | sw(f2, f3) | Z1 ~ X1 ", data=data)
-    with pytest.raises(MultiEstNotSupportedError):
+    with pytest.raises(NotImplementedError):
         feols(fml="Y  ~ 1 | csw(f2, f3) | Z1 ~ X1 ", data=data)
     # unsupported HC vcov
     with pytest.raises(VcovTypeNotSupportedError):
